refact: move relatorios to apps
