+++ bancodehoras/apps/relatorio/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class RelatorioConfig(AppConfig):
-    name = 'relatorio'
+    name = 'apps.relatorio'
