+++ bancodehoras/apps/relatorio/views.py
@@ -3,7 +3,7 @@ from django.shortcuts import render, redirect, HttpResponse
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.conf import settings
-from relatorio import controller
+from apps.relatorio import controller
 from datetime import datetime
 from apps.core.models import *
 from apps.movimentacao.views import seleciona_dados
